Fix warning

+++ src/tap/communication/sensors/encoder/wrapped_encoder.cpp
@@ -64,7 +64,7 @@ float WrappedEncoder::getVelocity() const
 void WrappedEncoder::alignWith(EncoderInterface* other)
 {
     tap::algorithms::WrappedFloat positionDifference = other->getPosition() - position;
-    float offset = positionDifference.getUnwrappedValue() / M_TWOPI * encoderResolution / gearRatio;
+    float offset = positionDifference.getUnwrappedValue() / static_cast<float>(M_TWOPI) * encoderResolution / gearRatio;
     this->encoderHomePosition += offset;
     this->encoder += offset;
     this->position = other->getPosition();
@@ -94,7 +94,7 @@ void WrappedEncoder::updateEncoderValue(uint32_t encoderActual)
     pastPosition = position;
     lastUpdateTime = tap::arch::clock::getTimeMicroseconds();
     position = tap::algorithms::Angle(
-        encoder.getUnwrappedValue() * M_TWOPI / encoderResolution * gearRatio);
+        encoder.getUnwrappedValue() * static_cast<float>(M_TWOPI) / encoderResolution * gearRatio);
 }
 }  // namespace encoder
